Звездочки теперь могут быть половинными

+++ app/src/main/java/com/skyfolk/quantoflife/db/EventDbEntity.kt
@@ -8,8 +8,8 @@ import java.util.*
 open class EventDbEntity(
     var quantId: String = "",
     var date: Long = 0,
-    var rate: Int? = null,
-    var numericValue: Int? = null,
+    var rate: Double? = null,
+    var numericValue: Double? = null,
     var note: String = "",
     @PrimaryKey
     var id: String = UUID.randomUUID().toString(),
